+++ src/client/MultipleRecipeView.java
@@ -10,38 +10,6 @@ import javafx.stage.Screen;
 import javafx.stage.Stage;
 import javafx.geometry.Pos;
 
-class Recipe extends HBox {
-    private Button recipeName;
-
-    Recipe() {
-        this.setPrefSize(500, 20); // sets size of recipe
-        this.setStyle("-fx-background-color: #DAE5EA; -fx-border-width: 0; -fx-font-weight: bold;"); // sets background
-
-        recipeName = new Button("Recipe Name");
-        recipeName.setPrefSize(Screen.getPrimary().getBounds().getWidth(), 20);
-        recipeName.setStyle("-fx-background-color: #DAE5FF; -fx-border-width: 0;");
-        recipeName.setPadding(new Insets(10, 10, 10, 10));
-        this.getChildren().add(recipeName);
-
-        addListeners();
-    }
-
-    public void addListeners() {
-        recipeName.setOnAction(e -> {
-            // go to detail recipe screen
-        });
-    }
-
-}
-
-class RecipeList extends VBox {
-    RecipeList() {
-        this.setSpacing(5); // sets spacing between recipe
-        this.setPrefSize(500, 560);
-        this.setStyle("-fx-background-color: #F0F8FF;");
-    }
-}
-
 public class MultipleRecipeView extends BorderPane {
     private Header header;
     private Footer footer;
