+++ src/client/RecipeList.java
@@ -1,0 +1,11 @@
+package client;
+
+import javafx.scene.layout.VBox;
+
+public class RecipeList extends VBox {
+  RecipeList() {
+    this.setSpacing(5); // sets spacing between recipe
+    this.setPrefSize(500, 560);
+    this.setStyle("-fx-background-color: #F0F8FF;");
+  }
+}
